dbsetuplambda/lambda/requirements.txt
boto3==1.24.12
botocore==1.27.12
jmespath==1.0.1
mysql-connector-python==8.0.29
protobuf==4.21.1
PyMySQL==1.0.2
python-dateutil==2.8.2
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.9
